refactor : quantity BigDecimal로 변경

+++ src/main/java/kongju/pickmeal/application/diet/data/DietDto.java
@@ -78,7 +78,7 @@ public class DietDto {
     public record IngredientsResponse(
             Long ingredientId,
             String name,
-            Double quantity,
+            BigDecimal quantity,
             IngredientUnit unit
     ) {
     }
